+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -3,15 +3,16 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.dto.order.OrderDto;
 import ru.yandex.practicum.dto.payment.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
 
     PaymentDto createPayment(OrderDto orderDto);
 
-    Double getTotalCost(OrderDto orderDto);
+    BigDecimal getTotalCost(OrderDto orderDto);
 
-    Double getProductCost(OrderDto orderDto);
+    BigDecimal getProductCost(OrderDto orderDto);
 
     void paymentSuccess(UUID paymentId);
